+++ src/widgets/button.h
@@ -72,7 +72,8 @@ BEGIN_C_DECLS
  * ```
  *
  * > 更多用法请参考：
- * [theme default](https://github.com/zlgopen/awtk/blob/master/demos/assets/raw/styles/default.xml#L31)
+ * [theme
+ * default](https://github.com/zlgopen/awtk/blob/master/demos/assets/raw/styles/default.xml#L31)
  *
  */
 typedef struct _button_t {